Some restructuring
- Partial fix for issue #15
