+++ src/dbxincluder/__init__.py
@@ -16,153 +16,15 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
 #
 
-"""Main module. Handles the docbook specific part."""
+"""Main module"""
 
-import base64
 import sys
 import lxml.etree
 
-from . import xinclude
-from .xinclude import DBXIException
+from . import docbook
+from . import utils
 
 __version__ = "0.1"
-
-
-def generate_id(elem):
-    """Generate a (per-document) unique ID for the XML element elem
-    :return: str"""
-    path = bytes(elem.getroottree().getpath(elem), encoding="utf-8")
-    return str(base64.urlsafe_b64encode(path), encoding="utf-8")
-
-
-def associate_new_ids(subtree):
-    """Assign elements their new ids as new attribute
-
-    :param subtree: The XIncluded subtree to process"""
-    idfixup = subtree.get("{http://docbook.org/ns/transclude}idfixup", "none")
-    if idfixup == "none":
-        return  # Nothing to do here
-
-    suffix = subtree.get("{http://docbook.org/ns/transclude}suffix")
-
-    if idfixup == "suffix":
-        assert len(suffix), "No suffix given"
-
-    for elem in subtree.iter():
-        old = elem.get("{http://www.w3.org/XML/1998/namespace}id")
-        if old is None:
-            continue
-
-        new = elem.get("{dbxincluder}newid", old)
-        if idfixup == "suffix":
-            new = old + suffix
-        elif idfixup == "auto":
-            new = old + "--" + generate_id(elem)
-        else:
-            # TODO: Get file and line from somewhere
-            raise DBXIException(elem, "idfixup type {0!r} not implemented".format(idfixup))
-
-        elem.set("{dbxincluder}newid", new)
-
-
-def find_target(elem, subtree, value, linkscope):
-    """Resolves reference to id value beginning from elem
-
-    :param elem: Source of reference
-    :param subtree: XIncluded subtree
-    :param value: ID of reference
-    :param linkscope: DB transclusion linkscope (local/near/global)
-    :return: Target element or None"""
-    if linkscope == "local":
-        target = subtree.xpath("./*[@xml:id={0!r}]".format(value))
-    elif linkscope == "near":
-        while elem.getparent() is not None:
-            elem = elem.getparent()
-            target = elem.xpath("./*[@xml:id={0!r}]".format(value))
-            if len(target):
-                return target[0]
-    elif linkscope == "global":
-        target = elem.xpath("//*[@xml:id={0!r}]".format(value))
-    else:
-        # TODO: Get file and line from somewhere
-        raise DBXIException(elem, "linkscope type {0!r} not implemented".format(linkscope))
-
-    return None if len(target) == 0 else target[0]
-
-
-def fixup_references(subtree):
-    """Fix all references if idfixup is set
-
-    :param subtree: subtree to process"""
-    idfixup = subtree.get("{http://docbook.org/ns/transclude}idfixup", "none")
-    if idfixup == "none":
-        return  # Nothing to do here
-
-    linkscope = subtree.get("{http://docbook.org/ns/transclude}linkscope", "near")
-    if linkscope == "user":
-        return  # Nothing to do here
-
-    for elem in subtree.iter():
-        if not elem.tag.startswith("{http://docbook.org/ns/docbook}"):
-            continue
-
-        idrefs = ["linkend", "linkends", "otherterm", "zone", "startref",
-                  "arearefs", "targetptr", "endterm"]
-        idrefs_multi = ["arearefs", "linkends", "zone"]
-        for attr, value in elem.items():
-            if attr not in idrefs:
-                continue
-
-            if attr in idrefs_multi:
-                assert False, "Not implemented"
-
-            # Find referenced element
-            target = find_target(elem, subtree, value, linkscope)
-
-            if target is None:
-                raise DBXIException(elem, "Could not resolve reference {0!r}".format(value))
-
-            # Update reference
-            new = target.get("{dbxincluder}newid")
-            if not new:
-                new = target.get("{http://www.w3.org/XML/1998/namespace}id")
-
-            elem.set(attr, new)
-
-
-def process_xml(xml, base_url=None, file=None):
-    """Same as xinclude.process_xml, but handle docbook attributes"""
-    if not isinstance(xml, bytes):
-        xml = bytes(xml, encoding="UTF-8")
-
-    tree = lxml.etree.fromstring(xml, base_url=base_url)
-
-    xinclude.process_tree(tree, base_url, file)
-    # Three passes:
-    # First, assign all elements a new ID
-    for subtree in tree.iter():
-        associate_new_ids(subtree)
-
-    # Second, fixup all references
-    for subtree in tree.iter():
-        fixup_references(subtree)
-
-    # Third, clean up our {dbxincluder} and the docbook transclude attributes
-    for elem in tree.iter():
-        newid = elem.get("{dbxincluder}newid")
-        if newid:
-            elem.set("{http://www.w3.org/XML/1998/namespace}id", newid)
-            del elem.attrib["{dbxincluder}newid"]
-
-        for name, _ in elem.items():
-            if name.startswith("{http://docbook.org/ns/transclude}"):
-                del elem.attrib[name]
-
-    # Remove unnecessary namespace declarations
-    lxml.etree.cleanup_namespaces(tree)
-
-    return lxml.etree.tostring(tree.getroottree(), encoding='unicode',
-                               pretty_print=True)
 
 
 def main(argv=None):
@@ -194,8 +56,8 @@ def main(argv=None):
         return 1
 
     try:
-        sys.stdout.write(process_xml(inputxml, base_url, path))
-    except xinclude.DBXIException as exc:
+        sys.stdout.write(docbook.process_xml(inputxml, base_url, path))
+    except utils.DBXIException as exc:
         sys.stderr.write(str(exc) + "\n")
         return 1
     except lxml.etree.XMLSyntaxError as exc:
